refactor(backoffice): split automation fragment routes
